+++ src/domain/cartItem.go
@@ -5,7 +5,7 @@ type CartItem struct {
 	*Item
 	Quantity int
 	InStock  bool
-	Total    float32
+	Total    float64
 }
 
 // 数量增加
@@ -17,6 +17,6 @@ func (ci *CartItem) IncrementQuantity() {
 // 计算总价: 单价 * 数量
 func (ci *CartItem) CalculateTotal() {
 	if ci.Item != nil && ci.Item.ListPrice != 0 {
-		ci.Total = ci.Item.ListPrice * float32(ci.Quantity)
+		ci.Total = ci.Item.ListPrice * float64(ci.Quantity)
 	}
 }
