+++ src/domain/item.go
@@ -3,8 +3,8 @@ package domain
 type Item struct {
 	ItemId     string  `db:"itemid"`
 	ProductId  string  `db:"productid"`
-	ListPrice  float32 `db:"listprice"`
-	UnitCost   float32 `db:"unitcost"`
+	ListPrice  float64 `db:"listprice"`
+	UnitCost   float64 `db:"unitcost"`
 	SupplierId int     `db:"supplier"`
 	Status     string  `db:"status"`
 	Attr1      string  `db:"attr1"`
